Added Admin class and account feature
I created an admin class. The idea is that they will be able to create teachers, students, etc. (features yet to come).

Both admins and teachers have usernames and encrypted passwords and are able to log on and log out. As of now, usernames and passwords cannot be changed. This feature will come at a later time.